Split ValidateCodeModal different usages in multiple components.
